enable `[sources]` section in Project.toml (#22)
So that everyone can instantiate dependency of expected versions by
running `Pkg.instantiate()`, and more importantly, to allow tests to
pass with custom versions of JET and JuliaLowering.

A drawback of this method is then we can't use packages installed in our
local package directory with simply running `pkg> dev JET JuliaLowering`.
The workaround is to temporarily comment out the `[sources]` section,
and then we can `pkg> dev JET JuliaLowering` and specify any checkout.
We'll need to use this method if we need to make changes to JET or JL
and test them against JETLS.

Another option is to use `git submodule`, but I think the solution with
`[sources]` is a bit simpler than managing `git submodule`, which is
known to be complex and tedious in general.