setup a special `AbstractAnalyzer` for JETLS
Currently, JETLS is using `JETAnalyzer`, which is the default analyzer
(`AbstractAnalyzer`) provided by JET.
However, `JETAnalyzer` is designed for interactive use within REPL in
mind, and thus when used in a language server, it presents the following
issues:
1. Too many false positives
2. Analysis performance

Regarding 1, there are the following aspects:
- Too many minor errors
- Unrelated errors: reports errors even in dependency packages that are
  not being actively developed by user

Regarding 2, the biggest bottleneck currently is that the analysis is
not performed incrementally, which is more of an issue with the bridge
between `AbstractAnalyzer` and top-level analysis rather than with
`AbstractAnalyzer` itself. Nevertheless, the performance of
`JETAnalyzer` may become an issue in the future, so it would be better
to setup a more lightweight analyzer for language server.

In this commit, I prepare `JETLSAnalyzer`, a dedicated analyzer for JETLS,
and use it for the full analysis in the language server.
Currently, `JETLSAnalyzer` only reports undefined global variables,
but of course, we will expand the types of errors reported in the future
while being careful not to increase false positive errors.