fix the warnings appearing in tests (#39)
Add a hack to prevent omitting empty `workspaceFolders` in the
serialization of `InitializeParams` during the `withserver` test
roundtrips, in order to fix the `"No workspaceFolders or rootUri in
InitializeRequest"` warnings appearing when `withserver` is called
without `workspaceFolders` nor `rootUri` specified.

`InitializeParams` is never serialized during the usual LS lifecycle,
so this hack doesn't affect any behavior of the LS for actual usecases.