disallow completion to be triggered by numeric character s
Some clients recognize numeric characters as "identifier characters" and
send completion requests when they are typed, but it feels very annoying
when the full completion list is displayed just by typing a numeric
character.
This commit fixes it so that completion isn't triggered in such cases.

Since we set `isIncomplete=false`, in cases where completion is already
happening, typing a numeric character will just properly filter the
completion list, not suppress the completion.