refactor server architecture to separate `Server` and `ServerState` (#44)
Define a new `Server` struct to make it easier to construct
`ServerState` in separation for testing purposes.
It also generally improves the separation between "the internal state"
of the server and "the server itself", which interacts with the external
world through the endpoint.
`Server` now holds `endpoint`, `callback`, and `state::ServerState` as
separate components.
Update all handlers to accept `Server` parameter and access `state`
through `server.state`.