LSP: correctly implement the `ResponseMessage.result` field (#48)
Currently, the `@interface` macro recognizes fields declared with a
default value of `= nothing` as "nullable". This works fine in most
cases, but causes issues when handling `ResponseMessage`.
```julia
@interface ResponseMessage @extends Message begin
    id::Union{Int, String, Nothing}
    result::Union{Any, Nothing} = nothing
    error::Union{ResponseError, Nothing} = nothing
end
```
According to the LSP specification, the `result` field must not be
omitted on success, and even when there's nothing specific to return
(like in `ShutdownResponse`), `null` must be explicitly returned.

Instead, `ResponseMessage.result` should not default to `nothing`, and
the `result` field in general should be explicitly specified in all
`ResponseMessage` instances.

This commit modifies the `@interface` macro to recognize fields as
"omittable" only when their declared type includes `Nothing`. It also
updates the definitions of `ResponseMessage` and other messages that
`@extends` it accordingly.