completion: specialize `@nospecialize`
`@nospecialize` is very common in our code base and ignoring it
completely
makes us unable to get completions for any methods using it,
since it makes method signatures completely broken.

Fortunately JL can handle it (probably because JL special-cases it),
so we can just keep it and support completions for such methods.