LSInterpreter: sync file cache to `textDocument/didSave` (#84)
Since aviatesk/JETLS.jl#73, the full analysis is synced to
`textDocument/didSave`, so the file cache used by `LSInterpreter`
also needs to be synced to that notification.

With this change, even if the user continues editing after saving,
as long as the last saved state is valid, the full analysis will
be performed appropriately.

The downside of this approach is that `ServerState` will manage
two separate file caches, which may introduce code complexity,
but that's probably unavoidable.

Related to this, in the future we will also need to properly update
`server.state.successfully_analyzed_file_infos` in response to
`textDocument/didChange` (for more accurate definition and completion),
but we'll leave that as a future task.