completion: simplify backslash completion implementation (#105)
After reconsidering the discussion at
https://github.com/aviatesk/JETLS.jl/pull/103#pullrequestreview-2943628486,
now I think it's better to have this completion even inside string
scopes. If the completion shows up, you can just ignore it, so it'd be
better than not having it available.

My previous comment was based on the fact that implementing general
completion features like `global_completions!` requires tokenization,
but for this specific backslash completion, I believe tokenization is
not really needed.