hover: Add local variable hover (#135)
* hover: Add local variable hover

Implement local_hover function to provide hover information for local
bindings by looking up their definitions and showing their source
locations in markdown format.

In the future, we would want to show the type of this local binding, but
that is only possible after the full JL integration. For now we'll just
show the location of the local binding

* define `select_target_binding_definitions` and use it from "definition"
and "hover"