Extract JSONRPC module into separate package (#121)
* Extract LSP module into separate package

Moves LSP definitions from src/LSP to a standalone package in LSP/
directory. Useful for other packages like JET to use the defintions in
the future (when LSP is registered as a general package).

Requires `Pkg.instantiate` after merge.

* Extract LSP module into separate package

Moves LSP definitions from src/LSP to a standalone package in LSP/
directory. Useful for other packages like JET to use the defintions in
the future (when LSP is registered as a general package).

Requires `Pkg.instantiate` after merge.

* Extract JSONRPC module into separate package

Moves JSONRPC functionality from src/JSONRPC/ to a standalone package in
JSONRPC/ directory. Updates the main package to depend on the new
JSONRPC package and introduces `LSEndpoint` wrapper for LSP-specific
method dispatching. JSONRPC is listed in `[sources]` section so it can
be used without registering it.