diagnostics: Report method errors
This commit is certainly quite useful as it does find new errors,
but it's still difficult to identify where the errors are coming from,
and it's still noisy in the current state.
In particular, it even reports type instabilities in the code generated
by `@testset`, so I'm not sure what to do about it.