inlay-hints: show `end module` suffix (#181)
As we begin full-scale use of JL in the future, we will make extensive
use of inlay-hints for features like inline type annotations. To build
the infrastructure for that purpose, I have implemented a simple
`end-module` suffix inlay-hint based on syntactic analysis.

For code patterns like: ```julia module A ... end ```

This automatically displays `#= module A =#` after the `end`. If there
is already a comment after `end`, the inlay hint is not displayed.