surface-analysis: fix issues with werid `@nospecialize` expand
Currently we remove all macro calls, but we treat `@nospecialize`
specially and keep it during lowering. However, for some reason there
are cases where `JS.first_byte` becomes 0 after `@nospecialize` is
applied, so we're adding that special case to surface-analysis.jl. This
is expected to be resolved by aviatesk/JETLS.jl#180.