use `Threads.@spawn` instead of `@async`
As suggested by the docstring of `@async`: > │ Warning > │ > │ It is
strongly encouraged to favor `Threads.@spawn` over `@async` always even
when no parallelism is required especially in > │ publicly distributed
libraries. This is because a use of `@async` disables the migration of
the parent task across worker > │ threads in the current implementation
of Julia. Thus, seemingly innocent use of `@async` in a library function
can have a large > │ impact on the performance of very different parts
of user applications.