full-analysis: Initial implementation of incremental analysis (#399)
This commit implements a prototype of Revise-based incremental
full-analysis. The idea is to delegate package state management to
Revise, similar to JET v0.11, enabling fast updates of analysis results
when incremental changes are made while avoiding issues like #357.

Although this implementation is incomplete, it works at least for JETLS
itself and its dependencies, making it possible to experiment with
various feature extensions that Revise-based incremental analysis
enables. For example, this approach makes it practical to cache type
inference results in `FileInfo` and update them on `DidChange` (though
memory leak risks return). This would make dynamic diagnostics, type on
hover, and inlay hints more feasible.

The following incomplete aspects need to be addressed for complete
migration to this new analysis approach:

- This mode is only enabled when `JETLS_DEV_MODE` is active.
  This is because JETLS currently loads Revise only when
  `JETLS_DEV_MODE` is enabled, and this feature relies on that Revise
  instance. However, there is no guarantee that existing analysis and
  Revise won't interfere with each other.
  I will enable this mode outside of `JETLS_DEV_MODE` only after I
  better understand Revise internals and can achieve finer-grained
  control over Revise's automatic file watching, allowing us to safely
  load Revise (or ReviseCore?) in a secure manner.

- Code loading is not yet implemented. This could be easily implemented
  using `Base.require`, but currently we use the approach of looking up
  `module_name` specified in `analysis_overrides` of initialization
  options from `Base.loaded_modules`. This means external packages
  cannot be analyzed with this new method - only packages already loaded
  as JETLS dependencies at server initialization.
  This is also related to the fact that this mode currently only works
  for package code. We need to reconsider how to load and analyze
  script code as well.

- The Revise-based tracking system needs adjustment for special packages
  like `Base` and `Compiler`.

- Generally, my understanding of Revise is still insufficient and the
  package monitoring approach may be immature.

Also checks in the changes for .JETLSConfig.toml to enable this new
analysis mode for JETLS itself.